# build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"PartnerServices": {
				"properties": {
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"price",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PartnerServicesDto": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"services": {
						"items": {
							"$ref": "#/components/schemas/PartnerServices"
						},
						"type": "array"
					}
				},
				"required": [
					"userId",
					"services"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PartnerProileDtoReq": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"location": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"description",
					"location",
					"rating",
					"image"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PartnerProfileDtoRes": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"location": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"registeredOn": {
						"type": "string",
						"format": "date-time"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"description",
					"location",
					"registeredOn",
					"rating",
					"image"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "crudapptest",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/services": {
			"get": {
				"operationId": "GetPartnerServiceDetailsById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PartnerServicesDto"
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/services/test": {
			"post": {
				"operationId": "Test",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "string"
							}
						}
					}
				}
			}
		},
		"/services/create": {
			"post": {
				"operationId": "CreatePartnerServices",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PartnerServicesDto"
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PartnerServicesDto"
							}
						}
					}
				}
			}
		},
		"/services/update": {
			"post": {
				"operationId": "UpdatePartnerServiceDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PartnerServicesDto"
							}
						}
					}
				}
			}
		},
		"/services/delete": {
			"delete": {
				"operationId": "DeletePartnerServices",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/partners": {
			"get": {
				"operationId": "GetPartnerDetailsById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PartnerProileDtoReq"
								}
							}
						}
					}
				},
				"tags": [
					"Partners"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/partners/test": {
			"post": {
				"operationId": "Test",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Partners"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "string"
							}
						}
					}
				}
			}
		},
		"/partners/create": {
			"post": {
				"operationId": "CreatePartner",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PartnerProfileDtoRes"
								}
							}
						}
					}
				},
				"tags": [
					"Partners"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PartnerProileDtoReq"
							}
						}
					}
				}
			}
		},
		"/partners/update": {
			"post": {
				"operationId": "UpdatePartnerDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Partners"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PartnerProileDtoReq"
							}
						}
					}
				}
			}
		},
		"/partners/delete": {
			"delete": {
				"operationId": "DeletePartner",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Partners"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}